Finish the rest of Chapter 15

+++ src/value.rs
@@ -2,7 +2,7 @@ use std::fmt::{Display, Formatter};
 use std::ops::{Add, Div, Mul, Neg, Sub};
 
 #[derive(Clone)]
-pub struct Value(pub f32);
+pub struct Value(pub f64);
 
 impl Display for Value {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
